Refactoring of river_reader.py file
This commit modifies the river_reader.py, adding the possibility
the read also some information about the biogeochemical variables
stored inside the rivers.xml file.

This commit adds a new function to river_reader.py file to read
the biogeochemical variables and stores the result of this
function in a global variable called BGC_VARS.

It also moved the code that reads the rivers inside a dedicated
function (but the global variable RIVERS is still available).

WARNING: this commit changes the way the tags "I" and "J" are read
from the xml file because it shifts the values by 1. In this way,
the indices written on the xml file are coherent with the indices
used by Fortran while the indices stored by the routines of the
river_Reader.py file are the indices used by Python (starting
from 0).