Added the possibility to have "partially defined" vars in rivers
When defining a river in the xml file, it is now possible to
specify only a subset of all the previous defined variables.
The variables that are not specified will be set to 0 inside
the RIVERS array. Moreover, the RIVERS array now implements
a method called `.get_variable_mask` that returns a mask M such
that RIVERS[M] contains only the mouths for which the variable
is defined