river_generation: Removing dedicated routines for Po
The variables related to the Po river has been implemented inside
the rivers.xml file in the previous commit. Therefore, there is
no reason to keep a separate routine for the Po river.

This commit removes that routine and ensures that the values of
each variable are saved only onto the cells related to the mouths
for which we actually have the data